Move `create-minimist-options.js`
